requirement.txt
api==0.0.7
asgiref==3.3.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
boto3==1.16.51
botocore==1.19.51
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
disqus==0.0.4
Django==3.1.3
django-admin-logs==1.0.1
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-excel==0.0.10
django-extensions==3.1.0
django-filter==2.4.0
django-humanize==0.1.2
django-import-export==2.5.0
django-mathfilters==1.0.0
django-sass-processor==0.8.2
django-storages==1.11.1
django-tables2==2.3.3
et-xmlfile==1.0.1
graphviz==0.16
humanize==3.2.0
idna==2.10
importlib-metadata==1.7.0
imutils==0.5.3
jdcal==1.4.1
jmespath==0.10.0
lml==0.1.0
MarkupPy==1.14
mysqlclient==2.0.2
nose==1.3.7
numpy==1.19.5
odfpy==1.4.1
opencv-python==4.5.1.48
openpyxl==3.0.5
pandas==1.2.0
paramiko==2.7.2
Pillow==8.0.1
pycparser==2.20
pycryptodomex==3.9.9
pydot3==1.0.9
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
pygraphviz==1.5
PyJWT==1.7.1
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzbar==0.1.8
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
storage==0.0.4.3
stream-django==1.6.2
stream-python==5.0.0
tablib==3.0.0
texttable==1.6.3
urllib3==1.26.2
whitenoise==5.2.0
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
